+++ lib/browser_of_babel.rb
@@ -2,6 +2,7 @@
 
 require_relative "browser_of_babel/library"
 require_relative "browser_of_babel/locator"
+require_relative "browser_of_babel/randomizer"
 require_relative "browser_of_babel/version"
 
 # Browser for Library of Babel.
@@ -10,8 +11,9 @@ module BrowserOfBabel
   # Base error class for BrowserOfBabel.
   class Error < StandardError; end
 
-  # Holotheca does not support this identifier.
+  # A holotheca does not support an identifier.
   class InvalidIdentifierError < Error; end
-  # Trying to go between library holotheca levels incorrectly, e.g. from hex to shelf.
+  # Trying to go between library holotheca levels incorrectly, e.g. from hex to shelf;
+  # or trying to extract text from a non-Page holotheca.
   class InvalidHolothecaError < Error; end
 end